Prepare for unifying HH autoimport behavior
Summary:
The typechecker and the compiler have different views of HH autoimporting—the typechecker thinks these names are global (and the compiler knows that they're actually in the HH namespace). This sets up the namespace resolution module to change the typechecker behavior piecemeal.

Unifying autoimport behavior is one of the blockers to eventually typecheck systemlib (which will allow us to kill HHIs).

Reviewed By: vassilmladenov

Differential Revision: D15656844

fbshipit-source-id: 6eeeaf5a4ebdca728e91ad13182dfc833033c33d