Add last_recheck_time to hh --json
Summary:
Refer to title, pasting description of original task for reference:

We use the init time to decide whether we need a saved state created at a given rev or not -- if the init time is very small then it's not worth it, we should move on and create one that has more impact instead. But, right now we have no structured way to get the init time from the hh_server after it has already started up, so the code is brittle. If we had some structured json that contained this value we could make this process much more robust.

Reviewed By: 2BitSalute

Differential Revision: D15594034

fbshipit-source-id: c91245a86e25b33c92c69d1a0c4991c6b650a87c