Obj_get_check should not report an error when the receiver object is ill-typed
Summary: The type `Terr` that we assign to ill-typed expressions should behave exactly like `Tany`.  However, because these are two separate constructors in the type ADT and because we do pattern matches on types, we occasionally fail to handle `Tany` and `Terr` consistently.  This diff fixes one such inconsistency, namely in dynamic property access check.  We currently error out on `$x->$y` if the type of `$x` is `Terr` but accept it if the type of `$x` is `Tany`.

Reviewed By: kmeht

Differential Revision: D15713215

fbshipit-source-id: cba8cc59cb00a93ff9c50c55e3bffea5e843a8c5