Namespace tree
Summary:
Build a tree data structure capable of storing information about namespaces.

Extend the SearchUtils enum list and other components of the autocomplete stack to contain information about namespaces.

The tree should contain a cumulative namespace name that can be used as well as individual segments of the namespace name.

This feature is not yet integrated into autocomplete.

Reviewed By: pittsw

Differential Revision: D15341603

fbshipit-source-id: 8e153514515268d1caa3d4250da88fe1fd15ea42