Add support for namespace aliases
Summary:
The namespace map system should support aliases that are configurable through .hhconfig.

An aliased namespace such as "Str\" might be a shortcut for "HH\Lib\Str\".

This is represented as a "shortcut" within the tree structure of namespaces.

This feature is not yet enabled for usage; that will follow later in this diffstack.

Reviewed By: pittsw

Differential Revision: D15696171

fbshipit-source-id: c6d1eb36278a9fa949b1b78c3ca51a3751cb3a6f