Add support for incremental generation of naming table saved state.
Summary: This change adds support for incremental saved state generation of the reverse naming table.  It works by keeping track of the original forward naming table that we loaded from the prior saved state, and when we generate the incremental saved state we diff that forward naming table with the current one, removing and adding entries in SQLite in order to get them to match.

Differential Revision: D15345594

fbshipit-source-id: 6e7efbc3fc39563675b98bbd20f1ed436e46dae3