Use member-ops for ElemArray
Summary:
We've got two copies of the ElemArray logic - one in the JIT, in `minstr-helpers`, and one in the interpreter, in `member-operations`. This diff uses the `member-operations` version in both places. Other IR ops, like ElemDict and ElemVec, work the same way.

There's one functional difference between the old and new logic: when the element is missing, the old JIT logic returned uninit while the new logic returned null. The new logic returns null in both cases. As a result of this change, we can restrict the destination types of these ops to be initialized. I would like to add the same restriction for the general ElemD / ElemU / ElemX ops as well, but they can still return uninit when the base is a class, so I'm leaving that as a follow-up.

Reviewed By: mxw

Differential Revision: D15521990

fbshipit-source-id: 12e05ae66db6c37a7017b9f5b44d91b7359a2e85