Add .hh_file=1 to all .hhas unit tests
Summary:
Add this temporarily so that we can kill the EnablePHP option.

This will be removed by D15662817. May not need to land this diff if D15662817
lands first, but that is currently blocked on introducing parse error for <?php
and I want to be able to get clean test runs for the next diffs on the stack.

Reviewed By: kmeht

Differential Revision: D15714916

fbshipit-source-id: 3e1a4a66aab5db1be2379d96db07646f6b059587