Move fake members info into per-continuation environment
Summary:
Currently the only information stored per-continuation (Next, Break, Catch, etc) is types for locals. But there are other environments that should be tracked per continuation, such as fake member validity, "using" variables, etc.

This diff is the first in a series that reorganizes our "continuation map" so that more information is stored per continuation. The first of these is the `fake_members` info - see the new tests for examples of how it's necessary to store it per continuation in order to be sound. In addition to control flow unsoundness, we also omitted to invalidate fake members when checking the body of a lambda.

At the same time I've taken the opportunity to refactor fake member logic, putting much of it into its own module `Typing_fake_members`. More to come.

Reviewed By: manzyuk

Differential Revision: D15621284

fbshipit-source-id: 2cc92c50781663f45e1736522ea39d5f1fb4e08d