Add error-case tests for Shapes::idx
Summary:
We'd like to use an optimized native implementation for Shapes::idx. This implementation should match the pure-PHP one both on valid inputs and in a variety of error cases, including:
- Type-hint errors: We should raise a fatal error including the argument index and actual and expected types.
- The DV array check: When the compat flags are on and an array is passed for argument 1, we should raise a notice and continue.

We split these tests out from the behavioral changes and check them in first so that we can confirm that the native implementation's behavior matches the old behavior.

Reviewed By: mxw

Differential Revision: D15699370

fbshipit-source-id: 5374a825edfe795524281b30e8255cdaae896f50