Fix a refcount bug in ClsRefGetTS
Summary:
In `ClsRefGetTS`, we have an input type structure list. Prior to this diff, in `addToReifiedGenericsTable`, this input will either get turned into static at which point its ref count will be decremented or nothing will happen. However, we do not want to discard the input if its refcount is 1 since first non param local still needs to point to it.
The goal of this diff is to make sure original input's refcount does not get consumed in `ClsRefGetTS`.

Reviewed By: markw65

Differential Revision: D15701929

fbshipit-source-id: 0c7b30d9068ff0b6181cc26a449e48c8361b93c3