Add FPushFuncRD as an optimizations for ReifiedName + FPushFunc
Summary:
This diff adds a new bytecode to optimize the combination of `ReifiedName` + `FPushFunc` since for reified generics we do know the name of the function statically but using `FPushFunc` was preventing us from using this static knowledge.
According to strobelight data, `ReifiedName` is a very expensive instruction. The goal is to remove this instruction completely.

Reviewed By: paulbiss

Differential Revision: D15624982

fbshipit-source-id: 29388fa421136582c031e74d2ee8e8d743046c4c