Reduce runtime of nzuncompress.php
Summary:
It was adding a million characters to a string, one character at a
time. In one-bit-refcount mode that forces a copy of the string on
each iteration, resuling in 1M string copies, with an average length
of 500k, and taking long enought to borderline timeout.

After constructing the string it nzcompresses it, and the rewrites the
first 4 bytes so that nzuncompress tries to create a string of length
2^32-1, and fatals.

There doesn't seem to be any need to create a 1M character string to
begin with, so make it 1000 chars instead.

Reviewed By: alexeyt

Differential Revision: D15730178

fbshipit-source-id: 1e836b94787f2ac9fca985729b2d3531316bc1ed