fix deadlock in connect thread
Summary:
So here's the deadlock: We get a connection request while we're closing a connection. We don't reset the connection flag until after the connection thread is done. So we send a rejection to the new client, and then try to take the socket transport lock again.

But, in the meantime, we switch back to the thread that's cleaning up and it calls StopConnectionThread. At this point, that thread has the socket transport lock, and StopConnectionThread tries to join the connection thread.

Reviewed By: ebluestein

Differential Revision: D15701465

fbshipit-source-id: 1415c7a270811a87d91dc03171b8382f60e0ba94