Clean inline stacks after retranslateAll
Summary:
We keep a side table mapping TC addresses to inline stacks that they're a part of, so that we can emit backtraces including inlined functions. After doing retranslateAll, we can remove TransProfile TC addresses from this map.

We spend quite a lot of time in production doing backtracing, both for exceptions and for other logging (exposed as the debug_backtrace helper). A lot of this backtrace time is in probing on this map, so it's helpful to clean it up.

This diff works by registering a call to `eraseInlineStacksInRange` at the point where we "delete" profiling translations. (To delete them, we just call madvise for the pages they're on.) To erase the stack, `filter_keys` copies over the hash table to a new table and treadmills the older one. Like all of the other mutating operations on TreadHashMap, `filter_keys` assumes there's only one thread writing to the table, which we enforce by holding the TC metadata lock.

I also changed the memory order with which we read the key here and in `acquireAndGrowIfNeeded`. I don't think a strong memory order is needed here, because of the single-writer assumption plus the fact that we always write the key with memory_order_release.

Reviewed By: markw65

Differential Revision: D15720928

fbshipit-source-id: 5ea5179955aff9691d0e4d7da4228cfc5ae44a2a