Fix issue with resolve
Summary:
In testing, the new sqlite autocomplete had a problem: it didn't display docblocks when you selected them in Nuclide.

Why did that happen?  Apparently the position information was absent, because of a value that was not being passed through correctly.

This diff looks up the correct positional information during sqlite autocomplete.

Reviewed By: arxanas

Differential Revision: D15713723

fbshipit-source-id: fe967d5eaf390aa1f1fa906335005937a4920e7b