Add decl type coercion target from return positions
Summary: As expressions are individually checked for coercion to a function's expected type, the decl type of the function hint needs to be threaded through to be considered for enforceability and dynamic coercion. In cases where we do not want to allow coercion, such as collection literals, we omit the decl type. This is abstracted away by the two make functions on the `ExpectedTy` module.

Reviewed By: Matt-Schellhas

Differential Revision: D15517735

fbshipit-source-id: 58a4f6e129f6aadfce802672e0a9e0b5eab3906f