Scan WeakRefDatas properly
Summary:
WeakRefDatas were not scanned properly.  They were only scanned properly if they were pointed to by WeakRefDataHandles.  When libxml was changed to use weakrefs, it used WeakRefData's directly and the GC would miss them, and not invalidate them properly.

This was the only fundamental thing preventing use of WeakRefData's directly, but there was no documentation or comments indicating the failure mode.

Reviewed By: ricklavoie

Differential Revision: D15748966

fbshipit-source-id: 0697bb5ecba4e7b9b986dd6a9eb6a4309b667cac