Remove Decl.mothed_pos
Summary:
[hack] Remove unnecesary use of Decl.method_pos

In order to add implicit `Naming_special_names.Classes.cStringish` interface to any class that have a __toString() method, the type checker looks for the method in the folded list of members which doesn't have Pos information, and then retrieve it from Decl_heap to obtain Pos information. However, this should only require looking in the shallow declaration which has Pos information within.

In this diff, we removed this unnecessary use of Decl.method_pos in Decl.class_decl function and replaced it with a direct access of Pos information from `shallow_class`.

Reviewed By: dlreeves

Differential Revision: D15616261

fbshipit-source-id: b7f7822c9c1b3b3f388e7c6dc96b04084cd59cc8