Add coercion dynamic ~> T when T is enforceable
Summary:
This diff allows coercion from dynamic to enforceable types. Currently, this is implemented for function return positions (D15517735). This leverages the `Type_test_hint_check` for is/as expressions to analyze the declared return type of a function and see if it is a valid coercion target for dynamic. The check is augmented in several places because `as` expressions support tests against types like shapes and tuples but HHVM does not enforce those types at function boundaries.

It should cause no new errors and shouldn't affect type inference, and we can add more allowed positions in subsequent diffs. Coercion is gated by an hhconfig flag

Reviewed By: dlreeves

Differential Revision: D15329223

fbshipit-source-id: 10ded08c91870f0bb656ca766f35c9314c9a2328