Add sequence intrinsic
Summary:
We'd like this (a no-op function that returns its last argument) for HHJS so we
can more easily control side effect ordering without adding more locals
everywhere.

This directs hackc to discard all but the last argument--the first few
expressions are eligible for DSE/DCE in the JIT so unless they have side
effects, they'll be discarded.

I would have implemented this as a builtin and a JIT optimization to eliminate
it in the JIT, but this is *much* easier and produces cleaner code anyways

Reviewed By: oulgen

Differential Revision: D15701910

fbshipit-source-id: 73fb4e01ca96e17306429a9883e33ab09ab2fadf