Allow an empty string for the SQLite filename in the saved state spec
Summary:
This change adds an option to skip loading the dependency table if it's not specified in the saved state spec.
I'm turning this option on by default because we currently always supply this value.

I intend to use this in the bulk type checker worker in order to skip sending/loading/using the dependency table.

Differential Revision: D13565393

fbshipit-source-id: 53b876097792122c59d9ef00ff814dfb17e7084b