Use separate queues for processing file change events and for responding to IDE queries
Summary:
We want to prioritize IDE queries at all costs. Service the IDE query queue first, and only then service the file changes queue.

This also changes the filtering logic to happen on the daemon side of things. Everything that passes data into it should be doing so dumbly, to reduce complexity.

Reviewed By: pittsw

Differential Revision: D15585283

fbshipit-source-id: 796c86fe2491bcb52c2d86c3aa57f0347ebe5a85