Support listening to file changes over the LSP with `workspace/didChangeWatchedFiles`
Summary:
This LSP notification lets us subscribe to file changes by leveraging the language client to provide us an implementation, so that we don't have to subscribe to Watchman ourselves.

This capability is only available via dynamic registration, so first I have to add support for dynamically registering capabilities in `clientLsp`. Once the language server is initialized, we send the language client a dynamic registration request to watch files for us (its response is `void`).

Reviewed By: 2BitSalute

Differential Revision: D15505773

fbshipit-source-id: 7f7e00c7e0b2d93d7f07b363b3460d88a9b98e94