Back out D15586692 and D15770624
Summary:
D15586692 was not yet ready, as it did not deal with varuous unit tests testing
Hack's support of genva(). Half of the tests on that diff were broken, but the
diff was landed anyway.

Back it out, together with D15770624, which tried to address some of these
tests, but not all of them.

fbshipit-source-id: caf76a2cbf6c41a44149298139c56eed3408ceb5