New si_fullitem data structure
Summary:
We need to know more information about symbols in order to make autocomplete more useful.

Rather than just knowing the kind, we need to know whether this particular symbol is valid in `acnew` / `acid` / `actype` contexts.  Then we can query for only those symbols that are valid in that specific context, which is less complicated than trying to figure this out at runtime.

We need to have a separate data structure for storing information about a symbol scan when compared to a search result.

This diff introduces new data structures in the symbol index builder to support this new information.

Reviewed By: pittsw

Differential Revision: D15745472

fbshipit-source-id: e0cc5e7d5a417e5c34067f4dce728b9c3c7d197d