Support TSimpleJSONProtocol decoding in json_decode
Summary:
TSimpleJSONProtocol is an encoding format used by thrift to serialize data containing strings in binary formats that would otherwise not be supported by JSON encoding.

The transformation is simple: whenever a non-printable character is detected, its binary value is hex-encoded, and stored in the last two digits of a \u escaped string. For example: `"\x80\x81"` is encoded as `"\u0080\u0081"`. This is crucially different from JSON for non-ASCII characters (`"\u{0080}"` and up), since normal JSON deserialization would treat this differently (see `utf16_to_utf8`) and decode it to `"\xc2\x80"`, whereas TSimpleJSON would decode it to `"\x80"`.

Reviewed By: billf

Differential Revision: D15571890

fbshipit-source-id: da0583bb121426c403b2afcb9a1ec3f23f3becfb