Relax constraints on closures in ExpressionParser::with_*
Summary:
The current signature prevents passing in closures that mutate
or move in the captured values.  For the former FnMut would suffice,
but if we want to allow captured values to be moved in, the closure
should be passed by move.  Therefore, move in a `F: FnOnce` closure
instead of passing it by reference.

Reviewed By: shiqicao

Differential Revision: D15773613

fbshipit-source-id: b3d57e2f6a938d4b645ce90cce334edb3b26681c