Add coercion to property types
Summary:
Obj_get takes a continuation that is threaded through its helper functions until a concrete object type is located, and the desired decl type (property, method, etc) is fetched from the environment. This is easiest to understand through the lens of assignment to object properties, `$x->a = $y`, so this diff implements exactly that.

This diff
- updates the continuation functions to carry the fetched declared type
- updates the expected record for bidirectional type checking of the right hand side of the assignment expression
- passes the declared type from the continuation to coercion

Reviewed By: kmeht

Differential Revision: D15739812

fbshipit-source-id: 5f5e0f3e71fb2413dc273177c9d690a9d4dab239