New inference: move arraykey index checks into type inference
Summary: Under a flag (`disallow-invalid-arraykey`) we check that the index expressions for map-like collection access and update are subtypes of `arraykey`. This *was* done in a post-inference TAST check. But subtype checks can influence type variables, and aid inference, so where possible we should do this in the type inference engine itself. This surfaces when attempting to check all lambdas using constraint-based inference.

Reviewed By: kmeht

Differential Revision: D15758920

fbshipit-source-id: f7319d3e0f483668dd1877895fac3599a82d8ecb