Produce an error when TAST contains unresolved type variables
Summary: At the moment we sometimes leak type variables into the typed AST. We would like to avoid that, so introduce a new disallow_unresolved_type_variables check to produce errors when we encounter unresolved type variables. By default the check is off and can be turned on with disallow_unresoved_type_variables=true in .hhconfig.

Reviewed By: manzyuk

Differential Revision: D15761506

fbshipit-source-id: 1e5a276f27c4fe73a2e2ea15d1b7653b50a60245