Set vasm block weights properly when splitting critical edges
Summary:
When splitting critical edges, we were always using the block weight
of the predecessor. This can be very wrong if the critical edge is
going from a cold bock to a hot one (the cold block gets the same
weight as the hot one). Since vasm-graph-color is increasingly relying
on block weights for its heuristic, this can cause it to make very bad
coloring decisions. Instead, use the smaller of the
successor/predecessor weights. This still isn't correct but it behaves
better. This is similar to what irgen does, except there it also
divides by two (I can't quite decide if this is good or not).

Also have vasm-print print out the weights when displaying a block, to
make this kind of issue more obvious.

Reviewed By: mxw

Differential Revision: D15812480

fbshipit-source-id: 57fb73911bc0bf58c0f7e027db9305b03a6bdea0