Use NewAnonymousClassName in rename_closure
Summary:
In sandbox mode, when generating anonymous class names at runtime, we
just stick a unique integer (generated by incrementing an atomic int)
on the end of each name.

In repo-authoritative mode, we may encode the class names into
AssertRAT opcodes, so we have to preserve the exact name in the
repo. When we flatten traits, we also copy their closures into the new
class, and give them a new name. Since we didn't use the atomic int
when we loaded the classes, we couldn't use it to give them a new
name; so we'd start with the id given to the original class, and keep
incrementing until we found an unused name. This seems reasonable
since the closure names are already more-or-less unique, except many
traits are used by hundreds of classes, so we ended up with O(n^2)
folly::format/makeStaticString/hash table lookups.

Instead, when reading a repo-authoritative repo, keep track of the
highest used index, so we can just continue from there when adding
more closures.

Reviewed By: binliu19

Differential Revision: D15825421

fbshipit-source-id: e968febe221faab3bafda56f2fd045181682e430