use local allocator in xml extensions
Summary:
Register the alloc/dealloc hook for libxml2 during process init.  This way, extensions (there are a few of them) based on libxml2 will allocate from the local allocator, because xml parsing is associated with a request.
This is safer than using the req-allocator, because even if some other random thread uses libxml2, it would still work.

Reviewed By: alexeyt

Differential Revision: D15553866

fbshipit-source-id: 5988695b6cbb5c3f85e8236a3ce1bf69b890f63b