change loadb to loadzx when the dest reg is used in wider width
Summary:
change loadb to loadzx when the dest reg is used in wider width
Use a simplify path after register allocation, since most of the opportunities
are created by register copies.
The post RA simplify has a separate entry point.
Also, introduce a new doPass1 in vasm-x64 that allows one addiional arg
to the pass, in addition to vunit. This streamlines the sequences of
invocations.
servicelabs flat.

Reviewed By: markw65

Differential Revision: D15378280

fbshipit-source-id: 3b4d14fc9328fd52e60b9a0e91d60d111dd0f23a