Use better hash for SrcKey's translation attempt count
Summary:
We were using the default hash in TBB on srckey's.  The default hash
is just a mutiplication.  Since our integers are not uniformly changing through
bit space, we need some kind of mixing hash.  This switches us to a mixing
version.

Play around with the pathology of the previous hash on godbolt:
https://godbolt.org/z/fVGVhF

We essentially end up using the same hash bucket for every srckey in a
function.

Reviewed By: binliu19

Differential Revision: D15843717

fbshipit-source-id: ab29a426152dc1056dbf1344bd4b4b451ba2bbd4