Typing_subtype.sub_string doesn't need to pattern-match on the type anymore
Summary: The function `Typing_subtype.sub_string` still uses the legacy approach based on pattern-matching to checking if a type is convertible to `string`.  This is only used to support the `allow_mixed` parameter, which is only used in `String_cast_check` to disable the error about `(string)` casts on an object if the object type is `mixed` or `nonnull`.  This special case can be easily handled separately.

Reviewed By: CatherineGasnier

Differential Revision: D15824770

fbshipit-source-id: 6a193cbc743fc54c34e2b83d3511da05ffe52f77