Better error message from Typing_subtype.sub_string
Summary:
`Typing_subtype.sub_string` uses a subtyping check with a union on the right, which introduces disjunctions and currently leads to rather poor error messages.  I'm working on improving error messages for disjunctions by showing errors generated in each branch, which leads to unwieldy error messages from `Typing_substring.sub_string` listing each of the supertypes tested against:
```
...
This is an arraykey
...
It is incompatible with Foo
...
This is a bool
...
It is incompatible with Foo
...
This is a float
...
It is incompatible with Foo
...
etc.
```
Let us wrap errors from `Typing_subtype.sub_string` with a more concise and informative error.

Reviewed By: andrewjkennedy, Wilfred

Differential Revision: D15824769

fbshipit-source-id: 5c01bc5a3f7ab2e06ebc7b8a7560c7aea18ce03e