Remove index for size of database
Summary:
In testing, a very counter-intuitive result: the size of the sqlite database appears to have critical relevance for performance.

When I first launched the sqlite database for symbols, it only had one index; later on, I added other indexes in the hope that they would optimize for other types of queries.

The opposite happened: As the size of the DB grew, performance of queries decreased.  In testing, removing the indexes put query performance back to where it should be (~20ms for most queries).

So we have two ways to shrink the size of the database:
* First, if we have a good candidate, we can add primary key items to some tables.  This prevents sqlite from adding primary keys of its own, which would increase storage size.
* Second, we can eliminate the unnecessary indexes.

Reviewed By: 2BitSalute

Differential Revision: D15864142

fbshipit-source-id: 2308a4442a9540c6efe31dda625301c2025dceda