Only run DCE in vasm-graph-color when rematerialization happened
Summary:
Rematerialization can create dead spills. We always generate the spill
because we don't necessarily know if all reloads will instead be
rematerialized. If they are, the spill is dead. We run a DCE pass
after spilling to remove these. We can be more efficient by only
running the DCE pass if we actually rematerialized something (there
shouldn't be dead code otherwise).

Reviewed By: mofarrell

Differential Revision: D15862985

fbshipit-source-id: 78f0c14671568165c224863cf82fb6eaeab255ee