Display full types when printing generic type mismatch (4110) errors
Summary:
Consider the following example:
```
<?hh // strict

function expect<T as int>(?T $_): void {}

function test<T>(?T $x): void {
  expect($x);
}
```
At the moment, we produce the following error message for it:
```
Typing[4110] Invalid argument
   --> /home/manzyuk/scratch/nullable_tparam.php
  3 | function expect<T as int>(?T $_): void {}
    |                           ^^ This is a nullable type
  5 | function test<T>(?T $x): void {
    |                  ^^ It is incompatible with a nullable type
  6 |   expect($x);
    |          ^^
```
"A nullable type is incompatible with a nullable type" is not a helpful (and is a very confusing) error message.  What is going on here?

To type check the function call `expect($x)`, we instantiate the generic parameter `T` in the definition of `expect` with a fresh type variable `#1`, which results in a constraint `#1 <: int` added to the type variable environment.  We also impose the constraint `?T <: ?#1` arising from the fact that the argument type must be a subtype of the declared parameter type.  The latter constraint simplifies to `T <: ?#1`, which is further simplified to a disjunction `(mixed <: ?#1) | (T <: #1)`.  (The type `mixed` is an omnipresent implicit upper bound on the generic parameter `T`.  There are two way in which the constraint `T <: ?#1` can be satisfied: through the upper bound `T <: mixed`, or through `#1 <: ?#1`.)  Both disjuncts are unsatisfiable (contradictory to the existing constraint `#1 <: int` on `#1`), so the disjunction is unsatisfiable too.  We don't display the errors arising from the attempts to satisfy the disjuncts and instead we generate a generic type mismatch error for the original types `?T` and `?#1`, and in the error message these types are rendered in a shallow way, as "a nullable type".  Usually this is OK: a subtyping constraint between two nullable types always simplifies to the case in which the subtype is nonnullable.  However, when further simplification of the constraint leads to a disjunction, we don't report mismatches between simplified types but between the original types.  In fact, changing that, so that when a disjunction is unsatisfiable we produce an error involving the types from which this disjunction was generated (`T` and `?#1` in this case) wouldn't be particularly helpful here either and would lead to an error message stating that a nullable type is incompatible with the generic parameter `T`.

I'm not yet sure how to properly fix this problem (I'm experimenting with a couple of ideas) but as a preliminary step I'm considering if displaying full types (rather than only textual representation of the root type constructors, e.g., "a nullable type", "a dict", "a tuple of length 3" etc.) would make error messages better.  While we are at it, we can also see if changing the wording makes the errors easier to understand.  Andrew complained to me that he didn't like "This is ... It is incompatible with ..." format.  In this diff, I've tried to change this to "Expected ... But got ..." (Perhaps "Expected a subtype of ... But got ..." would be even clearer.)  Some error messages require some massaging (rephrasing of typing reasons) to read nicely, but see if you like this format better.

With this change, the error for the above example reads as follows:
```
Typing[4110] Invalid argument
   --> /home/manzyuk/scratch/nullable_tparam.php
  3 | function expect<T as int>(?T $_): void {}
    |                           ^^ Expected ?[unresolved]
  5 | function test<T>(?T $x): void {
    |                  ^^ But got ?T
  6 |   expect($x);
    |          ^^
```
This is still not particularly illuminating but hopefully less confusing.  I'd like to have something like `Expected ?#1, where #1 is an instantiation of T and #1 <: int` or something like that.  Or perhaps a special error message, something like `T has no upper bounds compatible with int`.

Reviewed By: Wilfred

Differential Revision: D15781671

fbshipit-source-id: 6d20c5e07d00bd9dc30df94e673e2f5712810af4