Cleanup: remove fresh_tenv
Summary:
The `fun_def` type checking function uses `fresh_tenv` to defensively reset the environment. But there is no need (it's the only function that does this) because it creates a fresh environment at the top of the function with `EnvFromDef.fun_env tcopt f`.

Also remove some redundant code and exposed functions.

Reviewed By: manzyuk

Differential Revision: D15874762

fbshipit-source-id: cf91ad64cffc22b6c38b9ee2aa1f5f7b51f01c24