Refactor dispatch_call in typing
Summary: The `obj_get` function carries a continuation so that it can apply the appropriate continuation on the concrete member i.e. when the receiver is a union. Specifically, `dispatch_call` passes the `call` function, which creates typed expressions for function call arguments, to the `obj_get` continuation. However, these typed expressions only live within the scope of the continuation, so they are captured by mutable variables outside the call to `obj_get`. This idiom is repeated multiple times throughout `dispatch_call`. This diff cleans it up so that we have one helper that hides the mutable variables and just returns the results that get placed into the Tast node.

Reviewed By: manzyuk

Differential Revision: D15868340

fbshipit-source-id: 04776a60b5a6548f021d7a0418dddb13a6cba1a4