Add mode to hh_server to dump naming table saved state separately.
Summary:
Some new projects coming down the pipeline need the naming table saved state, but not the entire saved state.  This change adds a new argument to write out just the naming table.

Since the naming table only depends on parsing, we're also able to short circuit initialization, and write out the naming table saved state without doing a full typecheck.

I'm doing this as a new `hh_server` flag instead of making a separate tool because file hashes, which are a key component of the naming table saved state that we use to detect when files have changed, are based off the hash of the AST, not the hash of the file (this is because we keep around the AST longer than we do the text of the file).  In order to guarantee that the ASTs we produce are identical, we have to parse the files in exactly the same way, with exactly the same parser options.  Given that, it was easiest to convert `hh_server` to dump this information than to make a fully new tool.

Reviewed By: 2BitSalute

Differential Revision: D15814758

fbshipit-source-id: dd690298f6425fa822a1706a497a9e7165eb5a82