Shallowly check can_be_used_as_lvalue
Summary:
This removes the recursive checking of `can_be_used_as_lvalue`
via the `prefix_unary_expression_stack` mechanism that seems to
have been introduced to support parsing by-ref assignments
(no longer supported in HackC/HHVM).

*Note*: This logic is shaky because it depends on referential equality of nodes
in the stack, so it blocks the progress of migrating the parser to Rust.

Reviewed By: dabek

Differential Revision: D15787346

fbshipit-source-id: bd2cca9799be82496cfa3faba870eb377d3c667d