Group HH autoimport names by kind in the typechecker
Summary:
The compiler separates the autoimport names by whether they refer to a type, function, or constant. Before this diff, the typechecker had one big map for all autoimport names.

This diff updates the typechecker to separate the autoimport names to be consistent with the compiler. In practice, this should not be a behavioral change because no names are shared between the buckets (except the Hack array functions, which won't be updated to the HH namespace until next diff).

Reviewed By: vassilmladenov

Differential Revision: D15812654

fbshipit-source-id: 5a7eb41ff4f1efd1bdf2a1007d0063e053bd6159