Add coercion for static properties
Summary: This diff enables coercion from dynamic for assignment and initialization of static properties. It does 2 main things: thread back the declared type of the class member, and add a continuation to `class_get` inspired by the continuation on `obj_get`. This continuation a passthrough in most cases, but in the assignment case, it hands off the job of calling `Typing_coercion.coerce_type` to when the class member is actually fetched inside `class_get_`. This has the nice side benefit of finding each decl type in a situation where the "class" is actually a union of classname<(X|Y)>.

Reviewed By: Matt-Schellhas

Differential Revision: D15766951

fbshipit-source-id: f11375b60995cb73c2bdc427d8150108a80f8b7a